hl7-forecasting-app/requirements.txt
streamlit>=1.36.0
psycopg[binary]
plotly>=5.18.0
pandas>=2.1.0
numpy
databricks-sdk>=0.89.0
requests
